refactor: extract utility functions to their own files
