+++ export-user-data/functions/src/utils.ts
@@ -1,0 +1,3 @@
+export function replaceUID(path: string, uid: string) {
+  return path.replace(/{UID}/g, uid);
+}
